fix:doc and test
